Extract legacy overlay utilities
